refactor(desktop): organize workspace sidebar hooks
