+++ desktop/src/hooks/workspaces/derived/use-sidebar-shortcut-targets.ts
