+++ desktop/src/hooks/workspaces/derived/use-workspace-sidebar-activities.ts
