+++ desktop/src/hooks/workspaces/facade/use-sidebar-repo-group-state.test.tsx
