+++ desktop/src/hooks/workspaces/ui/use-workspace-sidebar-native-context-menu.ts
